fix: address review comment 4620252998 - share compose test helper
